portfolio-application/app/requirements.txt
Babel==2.14.0
blinker==1.7.0
click==8.1.7
dnspython==2.5.0
email-validator==2.1.0.post1
Flask==3.0.2
Flask-BabelEx==0.9.4
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-Principal==0.4.0
Flask-Security==3.0.0
Flask-Testing==0.8.1
Flask-WTF==1.2.1
gunicorn==21.2.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
mongomock==4.1.2
packaging==23.2
passlib==1.7.4
prometheus-client==0.19.0
pymongo==4.6.1
python-dotenv==1.0.1
sentinels==1.0.0
speaklater==1.3
Werkzeug==3.0.1
WTForms==3.1.2
